Implemented encryption.

+++ VirtualGuard.Runtime/Dynamic/Constants.cs
@@ -4,9 +4,9 @@ namespace VirtualGuard.Runtime.Dynamic
     public static class Constants
     {
         // reader
-        public static int RD_IV = 0; // dbg
-        public static int RD_HANDLER_ROT = 0;
-        public static int RD_BYTE_ROT = 0;
+        public static byte RD_IV = 0; // dbg
+        public static byte RD_HANDLER_ROT = 0;
+        public static byte RD_BYTE_ROT = 0;
 
         // data
         public static string DT_WATERMARK;
